Just put my first text edits into the readme file
Here is a much longer version of the note on the changes I described in the title above.

I am just trying to fill up this box a little so I can understand how the information is stored and displayed for future reference.  Especially since I am about to try and teach my daughters to code - I should go through all the steps and fill in all the boxes first.